+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## [3.2.0]
+
+- Add Advisor model tool to help when Cline hits a roadblock (available with OpenRouter and Anthropic)
+- Add new Task/Chat mode toggle to turn Cline into a conversational partner, rather than a task-completing agent
+- Easily switch between API providers and models using a new popup menu under the chat field
+- Add VS Code LM API provider to run models provided by other VS Code extensions (e.g. GitHub Copilot). Shoutout to @julesmons, @RaySinner, and @MrUbens for putting this together!
+- Add on/off toggle for MCP servers to disable them when not in use. Thanks @MrUbens!
+- Add Auto-approve option for individual tools in MCP servers. Thanks @MrUbens!
+
 ## [3.1.10]
 
 - New icon!